+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Create a API Docs.
+- Add the optimistic minicart strategy props.
 
 ## [2.10.6] - 2019-02-12
 ### Fixed